// apps/api/src/generated/prisma/enums.ts
export const CashierApprovalStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  REVOKED: 'REVOKED'
} as const

export type CashierApprovalStatus = (typeof CashierApprovalStatus)[keyof typeof CashierApprovalStatus]


export const SubscriptionStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED'
} as const

export type SubscriptionStatus = (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus]


export const AssignmentStartReason = {
  INVITATION: 'INVITATION',
  CLIENT_BLOCKED_CASHIER: 'CLIENT_BLOCKED_CASHIER',
  CLIENT_REPORTED_CASHIER: 'CLIENT_REPORTED_CASHIER',
  CASHIER_BLOCKED_CLIENT: 'CASHIER_BLOCKED_CLIENT',
  CASHIER_UNAVAILABLE: 'CASHIER_UNAVAILABLE',
  ADMINISTRATIVE: 'ADMINISTRATIVE'
} as const

export type AssignmentStartReason = (typeof AssignmentStartReason)[keyof typeof AssignmentStartReason]


export const AssignmentEndReason = {
  CLIENT_BLOCKED_CASHIER: 'CLIENT_BLOCKED_CASHIER',
  CLIENT_REPORTED_CASHIER: 'CLIENT_REPORTED_CASHIER',
  CASHIER_BLOCKED_CLIENT: 'CASHIER_BLOCKED_CLIENT',
  CASHIER_UNAVAILABLE: 'CASHIER_UNAVAILABLE',
  CLIENT_SUSPENDED: 'CLIENT_SUSPENDED',
  CASHIER_SUSPENDED: 'CASHIER_SUSPENDED',
  ACCOUNT_DELETED: 'ACCOUNT_DELETED',
  ADMINISTRATIVE: 'ADMINISTRATIVE'
} as const

export type AssignmentEndReason = (typeof AssignmentEndReason)[keyof typeof AssignmentEndReason]


export const ReassignmentReason = {
  CLIENT_BLOCKED_CASHIER: 'CLIENT_BLOCKED_CASHIER',
  CLIENT_REPORTED_CASHIER: 'CLIENT_REPORTED_CASHIER',
  CASHIER_BLOCKED_CLIENT: 'CASHIER_BLOCKED_CLIENT',
  CASHIER_UNAVAILABLE: 'CASHIER_UNAVAILABLE',
  ADMINISTRATIVE: 'ADMINISTRATIVE'
} as const

export type ReassignmentReason = (typeof ReassignmentReason)[keyof typeof ReassignmentReason]


export const ReassignmentStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type ReassignmentStatus = (typeof ReassignmentStatus)[keyof typeof ReassignmentStatus]


export const ConversationStatus = {
  ACTIVE: 'ACTIVE',
  CLOSED: 'CLOSED'
} as const

export type ConversationStatus = (typeof ConversationStatus)[keyof typeof ConversationStatus]